an embarassing one
a parameter that was of type 'float' was being dropped because the set_value method from the protocol wizard was written such that if a 'type' didn't match something in its if/elif chain it silently dropped it. 'float' was not in its chain. that's really bad. try to catch the value if we can and it should throw an error if you can't.